remove inherited class and move it to original class
